+++ simulator/runtime/rdo_logic_dptsome.cpp
@@ -13,3 +13,19 @@
 // ----------------------------------------------------------------------- SYNOPSIS
 #include "simulator/runtime/rdo_logic_dptsome.h"
 // --------------------------------------------------------------------------------
+
+OPEN_RDO_RUNTIME_NAMESPACE
+
+// --------------------------------------------------------------------------------
+// -------------------- RDODPTSome
+// --------------------------------------------------------------------------------
+RDODPTSome::RDODPTSome(const LPRDORuntime& pRuntime, LPIBaseOperationContainer parent)
+	: RDOLogicSimple(pRuntime, parent)
+{
+	pRuntime->getFreeDPTId();
+}
+
+RDODPTSome::~RDODPTSome()
+{}
+
+CLOSE_RDO_RUNTIME_NAMESPACE
