+++ simulator/runtime/rdo_logic_dptsome.h
@@ -37,6 +37,4 @@ private:
 
 CLOSE_RDO_RUNTIME_NAMESPACE
 
-#include "simulator/runtime/rdo_logic_dptsome-inl.h"
-
 #endif // _LIB_RUNTIME_LOGIC_DPTSOME_H_
